Enable browser caching of signed assets
When asset signing is enabled (i.e. when assets are referenced
in the front-end with the deployment version in the URL, like
`/static/versionXXXXXXX/...`), it is safe to indicate to the
browser that it can "forever cache" the assets.

We add a `Cache-Control: max-age=31536000` header for that purpose,
if we found out that the original URL was signed.