Add store code to CountryMapping list to identify store
When selecting a new mapping for a country, only the store name was shown, as this can be the language to select the list can be filled with a lot of the same names making it hard to select the correct store. By adding the storecode to the label its easier to select the correct store.